Added Section Called "Completed Items"
Implemented a new function: all checked items will be moved to a new section called "Completed Items".
A bug: the "Reset List" button doesn't reset the "Completed Items" list.
For style: would be nice to have the "To-Do List" section on the left, and the "Completed Items" section on the right.